+++ src/models/user_activation_document.ts
@@ -1,7 +1,7 @@
 import { Container } from "@azure/cosmos";
 import { CosmosdbModel } from "@pagopa/io-functions-commons/dist/src/utils/cosmosdb_model";
 import { IoCourtesyDigitalAddressActivation } from "../generated/IoCourtesyDigitalAddressActivation";
-import { NewActivationDocument, RetrievedActivationDocument } from "../types";
+import { NewActivationDocument, RetrievedActivationDocument } from "../utils/types";
 
 export class UserActivationDocument extends CosmosdbModel<
   IoCourtesyDigitalAddressActivation,
