+++ src/utils/config.ts
@@ -1,8 +1,17 @@
 import { pipe } from "fp-ts/function";
 import * as E from "fp-ts/lib/Either";
-import { Errors } from "io-ts";
-import { Config } from "../types";
+import type { Errors } from "io-ts";
+import * as t from "io-ts";
 import { log } from "./logger";
+
+export const Config = t.type({
+  COSMOSDB_KEY: t.string,
+  COSMOSDB_NAME: t.string,
+  COSMOSDB_URI: t.string,
+  SERVER_PORT: t.string
+});
+
+export type Config = t.TypeOf<typeof Config>;
 
 export const getConfig = (env: NodeJS.ProcessEnv): E.Either<Errors, Config> =>
   pipe(env, Config.decode);
